+++ docs/fr_FR/changelog.md
@@ -4,6 +4,10 @@
 >
 >S'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte.
 
+# 22/02/2023
+
+- Optimisation du calcul : Eviter si non nécessaire (lampe éteinte et pas d'historisation)
+
 # 21/02/2023
 
 - Refactorisation du code